+++ src/onegov.event/HISTORY.rst
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ~~~~~~~~~~
+
+- Adds onegov.search integration.
+  [href]
+
 0.0.5 (2015-09-15)
 ~~~~~~~~~~~~~~~~~~~